+++ commands/main/support.js
@@ -11,7 +11,9 @@ class Support extends Command {
     const inviteEmbed = this.MessageEmbed()
       .setColor('#8fde5d')
       .setTitle('CatBot Support')
-      .setDescription('[Join the Discord](https://discord.gg/p5GRCSh)')
+      .setDescription(
+        `[Join the Discord](${client.config['bot']['support_server']})`
+      )
       .setTimestamp()
       .setFooter('Support Link Request', client.user.avatarURL());