+++ commands/mhgu/hzv.js
@@ -6,6 +6,9 @@ const Utils = require('../../bot/utils');
 const imageJSON = require('../../source_files/MonsterDataImages/mhgu_monster_map.json');
 const imageMap = Utils.getDataAsMap(imageJSON);
 
+const iconJSON = require('../../source_files/MonsterDataImages/mhgu_monster_icon_map.json');
+const iconMap = Utils.getDataAsMap(iconJSON);
+
 class Hzv extends Command {
   constructor() {
     super('hzv', 'hzv [monster name]', 'Get hzv info for a specific monster');
@@ -13,22 +16,27 @@ class Hzv extends Command {
 
   monsterEmbed(message, name, rawEmbed = this.MessageEmbed, menu = this.menu) {
     const image = imageMap.get(name);
+    const icon = iconMap.get(name);
 
     logger.debug('hzv log', { type: 'hzvRead', name: name });
 
-    const filename = image.fileName
-      .split("'")
-      .join('')
-      .split('-')
-      .join('')
-      .split(' ')
-      .join('');
+    const thumbnail = new MessageAttachment(
+      icon.imagePath,
+      icon.fileName.replace(/[',\s,-]/g, '')
+    );
+    const embedImage = new MessageAttachment(
+      image.imagePath,
+      image.fileName.replace(/[',\s,-]/g, '')
+    );
+
+    const attachUrl = name => `attachment://${name}`;
 
     const embed = rawEmbed()
       .setColor('#8fde5d')
       .setTitle(`__**${image.title}**__`)
-      .attachFiles([new MessageAttachment(image.imagePath, filename)])
-      .setImage(`attachment://${filename}`)
+      .attachFiles([embedImage, thumbnail])
+      .setThumbnail(attachUrl(thumbnail.name))
+      .setImage(attachUrl(embedImage.name))
       .setTimestamp()
       .setFooter(image.title);
 
@@ -39,7 +47,7 @@ class Hzv extends Command {
     let input = args.join('').toLowerCase();
 
     if (imageMap == null) {
-      return message.channel.send(this.serverErrorEmbed());
+      return message.channel.send(await this.serverErrorEmbed());
     }
 
     if (!imageMap.has(input)) {