+++ commands/mhw/weapon.js
@@ -72,7 +72,7 @@ class Weapon extends Command {
     let input = args.join('').toLowerCase();
 
     if (client.mhwWeapons == null) {
-      return message.channel.send(this.serverErrorEmbed());
+      return message.channel.send(await this.serverErrorEmbed());
     }
 
     if (!client.mhwWeapons.has(input)) {