+++ source_files/MonsterDataImages/main.js
@@ -2,11 +2,16 @@ const fs = require('fs');
 const { parse } = require('path');
 const mhwFolder = './source_files/MonsterDataImages/assets/mhw/monster/';
 const mhguFolder = './source_files/MonsterDataImages/assets/mhgu/monster/';
+const mhguIconFolder =
+  './source_files/MonsterDataImages/assets/mhgu/monster_icons/';
 
 const mhwFileNames = [];
 const mhwFileObjects = {};
 const mhguFileNames = [];
 const mhguFileObjects = {};
+
+const mhguIconFileNames = [];
+const mhguIconFileObjects = {};
 
 fs.readdirSync(mhwFolder).forEach(file => {
   mhwFileNames.push(file);
@@ -16,13 +21,22 @@ fs.readdirSync(mhguFolder).forEach(file => {
   mhguFileNames.push(file);
 });
 
+fs.readdirSync(mhguIconFolder).forEach(file => {
+  mhguIconFileNames.push(file);
+});
+
 fill(mhwFileNames, mhwFileObjects, mhwFolder);
 fill(mhguFileNames, mhguFileObjects, mhguFolder);
+fill(mhguIconFileNames, mhguIconFileObjects, mhguIconFolder);
 
 write('./source_files/MonsterDataImages/mhw_monster_map.json', mhwFileObjects);
 write(
   './source_files/MonsterDataImages/mhgu_monster_map.json',
   mhguFileObjects
+);
+write(
+  './source_files/MonsterDataImages/mhgu_monster_icon_map.json',
+  mhguIconFileObjects
 );
 
 function replaceAll(string, searchString, replaceString) {